Source nodes, node events, more node types, preset work, selection menu

+++ VRCOSC.App/SDK/Modules/Heartrate/HeartrateModule.cs
@@ -213,7 +213,7 @@ public abstract class HeartrateModule<T> : Module where T : HeartrateProvider
     {
         if (GetSettingValue<bool>(HeartrateSetting.SmoothValue))
         {
-            currentValue = Interpolation.DampContinuously(currentValue, targetValue, GetSettingValue<int>(HeartrateSetting.SmoothValueLength) / 2f, 50f);
+            currentValue = (float)Interpolation.DampContinuously(currentValue, targetValue, GetSettingValue<int>(HeartrateSetting.SmoothValueLength) / 2f, 50f);
         }
         else
         {
@@ -242,7 +242,7 @@ public abstract class HeartrateModule<T> : Module where T : HeartrateProvider
 
         if (GetSettingValue<bool>(HeartrateSetting.SmoothAverage))
         {
-            currentAverage = Interpolation.DampContinuously(currentAverage, targetAverage, GetSettingValue<int>(HeartrateSetting.SmoothAverageLength) / 2f, 50f);
+            currentAverage = (float)Interpolation.DampContinuously(currentAverage, targetAverage, GetSettingValue<int>(HeartrateSetting.SmoothAverageLength) / 2f, 50f);
         }
         else
         {
