+++ VRCOSC.App/Utils/Interpolation.cs
@@ -7,13 +7,13 @@ namespace VRCOSC.App.Utils;
 
 public static class Interpolation
 {
-    public static float Lerp(float start, float final, float amount) => start + (final - start) * amount;
+    public static double Lerp(double start, double final, double amount) => start + (final - start) * amount;
 
-    public static float DampContinuously(float current, float target, float halfTimeMilli, float elapsedTimeMilli)
+    public static double DampContinuously(double current, double target, float halfTimeMilli, float elapsedTimeMilli)
     {
         var exponent = elapsedTimeMilli / halfTimeMilli;
         return Lerp(current, target, 1 - MathF.Pow(0.5f, exponent));
     }
 
-    public static float Map(float source, float sMin, float sMax, float dMin, float dMax) => dMin + (dMax - dMin) * ((source - sMin) / (sMax - sMin));
+    public static double Map(double source, double sMin, double sMax, double dMin, double dMax) => dMin + (dMax - dMin) * ((source - sMin) / (sMax - sMin));
 }
